Relocate *_index_sorts; delete tabs/{locations,names}_helper.rb
`locations_index_sorts` and `names_index_sorts` were the only
methods left in `tabs/locations_helper.rb` and `tabs/names_helper.rb`
— both non-tab utilities (sort options for `add_sorter`). Move
them to the corresponding `app/helpers/{locations,names}_helper.rb`
and delete the now-empty tabs/* files. Rails auto-helper inclusion
keeps the methods reachable from every consuming ERB without further
wiring.

Tidy up dangling `# Replaces Tabs::{Locations,Names}Helper#...`
docstring lines in 5 Tab POROs whose original helper-method
counterparts no longer exist.
