+++ app/helpers/locations_helper.rb
@@ -39,4 +39,18 @@ module LocationsHelper
 
     SpeciesList.safe_find(species_lists[0])
   end
+
+  # Sort options passed to `add_sorter` from the Locations index.
+  # When the query is currently ordered by rss_log, "Updated" maps
+  # to the rss_log timestamp instead of the location's updated_at.
+  def locations_index_sorts(query: nil)
+    rss_log = query&.params&.dig(:order_by) == :rss_log
+    [
+      ["name", :sort_by_name.t],
+      ["created_at", :sort_by_created_at.t],
+      [(rss_log ? "rss_log" : "updated_at"), :sort_by_updated_at.t],
+      ["num_views", :sort_by_num_views.t],
+      ["box_area", :sort_by_box_area.t]
+    ]
+  end
 end
