+++ app/helpers/names_helper.rb
@@ -194,4 +194,17 @@ module NamesHelper
               form_to_inherit_classification_of_name_path(name.id))
     end
   end
+
+  # Sort options passed to `add_sorter` from the Names index. When
+  # the query is currently ordered by rss_log, "Updated" maps to
+  # the rss_log timestamp instead of the name's updated_at.
+  def names_index_sorts(query: nil)
+    rss_log = query&.params&.dig(:order_by) == :rss_log
+    [
+      ["name", :sort_by_name.t],
+      ["created_at", :sort_by_created_at.t],
+      [(rss_log ? "rss_log" : "updated_at"), :sort_by_updated_at.t],
+      ["num_views", :sort_by_num_views.t]
+    ]
+  end
 end
